+++ src/components/MyUpdates/MyUpdates.module.scss
@@ -1,10 +1,10 @@
 @import 'styles/_global.scss';
 
-.MijnUpdates > ul {
+.MyUpdates > ul {
   @include reset-list-style;
 }
 
-.MijnUpdateItem {
+.MyUpdateItem {
   border-top: 1px solid $color-neutral-grey2;
   padding-top: 2rem;
   padding-bottom: 2rem;
