+++ src/hooks/mijn-updates-api.hook.js
@@ -2,5 +2,5 @@ import { ApiUrls } from 'App.constants';
 import paginatedApiHook from './paginated-api.hook';
 
 export default (offset, limit) => {
-  return paginatedApiHook(ApiUrls.MIJN_UPDATES, offset, limit);
+  return paginatedApiHook(ApiUrls.MY_UPDATES, offset, limit);
 };
