+++ src/pages/MyUpdates/MyUpdates.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import MijnUpdates from './MijnUpdates';
+import MyUpdates from './MyUpdates';
 import AppState from 'AppState';
 
 const appState = { MY_UPDATES: {} };
@@ -8,7 +8,7 @@ const appState = { MY_UPDATES: {} };
 it('Renders without crashing', () => {
   shallow(
     <AppState value={appState}>
-      <MijnUpdates />
+      <MyUpdates />
     </AppState>
   );
 });
